requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.0
black==23.3.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.8
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
wsproto==1.2.0
yarl==1.9.2
